Responsive styling, more restructuring, added favicon
