+++ src/containers/HomePage/HomePage.css
@@ -1,4 +1,6 @@
 .HomePage {
+    width: 80%;
+    margin-left: 10%;
     grid-column: 2 / 3;
     grid-row: 2;
 }
@@ -24,7 +26,7 @@
     }
 }
 
-@media (min-width: 576px) and (max-width: 767.98px) {
+@media (min-width: 576px) and (max-width: 812.98px) {
     .Content {
         padding: 1rem 2rem;
     }
