+++ src/containers/LoginReg/LoginReg.css
@@ -1,0 +1,77 @@
+.Login, .Register {
+    width: 80%;
+    margin-left: 10%;
+    text-align: center;
+    grid-column: 3 / 4;
+    grid-row: 2;
+}
+
+.Content {
+    background-color: rgba(255, 255, 255, 0.5);
+    border-radius: 10px;
+    padding: 1rem;
+    margin: 5rem 0 0 0;
+}
+
+.Logo {
+    width: 100%;
+    margin: 1rem 0 0.5rem 0;
+}
+
+.Content h3 {
+    color: #791313;
+    font-weight: bold;
+    margin: 1rem 0;
+}
+
+.Content input {
+    box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
+    border: none;
+    margin-bottom: 0.5rem;
+    width: 80%;
+}
+
+.Content .SubmitButton {
+    margin-top: 0.5rem;
+    margin-left: 60%;
+    width: 30%;
+    border-radius: 4px;
+    background-color: #25116c;
+    color: #ffffff;
+    font-weight: bold;
+}
+
+/* Responsive Styling */
+@media (max-width: 575.98px) {
+    .Content {
+        margin: 3rem 0 0 0;
+    }
+
+    .Logo {
+        width: 65%;
+    }
+}
+
+@media (min-width: 576px) and (max-width: 812.98px) {
+    .Content {
+        margin-top: 2rem;
+    }
+
+    .Logo {
+        width: 55%;
+    }
+}
+
+@media (max-width: 812.98px) {
+    .Login, .Register {
+        grid-column: 2 / 3;
+    }
+
+    .Logo {
+        margin: 0.5rem 0 0.25rem 0;
+    }
+
+    .Content h3 {
+        margin-top: 0;
+    }
+}
