+++ src/containers/LoginReg/Register.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import classes from "./Register.css";
+import classes from "./LoginReg.css";
 import axios from "axios";
 
 class RegisterPage extends Component {
@@ -7,21 +7,20 @@ class RegisterPage extends Component {
         return (
             <div className={classes.Register}>
                 <div className={classes.Content}>
-                    <div className={classes.ImageHeader}>
-                        <img
-                            src={require("../../assets/images/logo03.png")}
-                            alt="Campus Navigator Logo"
-                        />
-                        <h3>Sign Up</h3>
-                    </div>
+                    <img
+                        className={classes.Logo}
+                        src={require("../../assets/images/logo03-2.png")}
+                        alt="Campus Navigator Logo"
+                    />
+                    <h3>Sign Up</h3>
                     <form>
                         <input placeholder="Create Username" />
                         <input placeholder="Create Password" />
                         <input placeholder="Confirm Password" />
-                        <div
+                        {/*<div
                             className="g-recaptcha"
                             data-sitekey="6Lc8MHUUAAAAAD8YIzicLlA-pa-8dTN4qFycOLoU"
-                        />
+                        />*/}
                         <input
                             type="submit"
                             className={classes.SubmitButton}
