--- src/containers/Register/Register.css
@@ -1,46 +1,0 @@
-.Register {
-    width: 80%;
-    margin-left: 10%;
-    text-align: center;
-    grid-column: 3 / 4;
-    grid-row: 2;
-}
-
-.Content {
-    background-color: rgba(255, 255, 255, 0.5);
-    border-radius: 10px;
-    padding: 1rem;
-}
-
-img {
-    width: 100%;
-}
-
-input {
-    box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
-    margin: 0.3rem;
-    border: none;
-    width: 80%;
-}
-
-.SubmitButton {
-    margin-left: 60%;
-    width: 30%;
-    border-radius: 4px;
-    background-color: #25116c;
-    color: #ffffff;
-    font-weight: bold;
-}
-
-h3 {
-    color: #791313;
-    font-weight: bold;
-    padding-bottom: 0px;
-    margin-bottom: 0px;
-}
-
-@media (max-width: 767.98px) {
-    .Register {
-        grid-column: 2 / 3;
-    }
-}
