+++ src/index.css
@@ -1,4 +1,5 @@
 html {
+  background-color: #BBB;
   font-size: 16px;
 }
